feat(*): add external support for method

+++ method.go
@@ -14,7 +14,7 @@ var _ Doc = (*MethodDoc)(nil)
 // MethodDoc .
 type MethodDoc struct {
 	BasicDoc
-	Parent string `json:"parent"`
+	Extra []byte `json:"extra"` // for further usage
 }
 
 // Marshal .
@@ -32,25 +32,14 @@ func (md *MethodDoc) GetID() DID {
 	return md.ID
 }
 
-var _ MethodManager = (*MethodRegistry)(nil)
-
-// MethodRegistry .
-type MethodRegistry struct {
-	config *MethodConfig
-	table  RegistryTable
-	docdb  DocDB
-	logger logrus.FieldLogger
-	admins []DID // admins of the registry
-	// network
-}
-
 var _ TableItem = (*MethodItem)(nil)
 
-// MethodItem reperesents a method item, element of registry table.
+// MethodItem reperesents a method item, element of registry table,
+// it stores all data about a did.
 // Registry table is used together with docdb.
 type MethodItem struct {
 	BasicItem
-	Owner DID // owner of the method, is a did
+	Owner DID // owner of the method, is a did, TODO: owner ==> owners
 }
 
 // Marshal .
@@ -68,104 +57,120 @@ func (mi *MethodItem) GetID() DID {
 	return mi.ID
 }
 
-// NewMethodRegistry news a MethodRegistry
-func NewMethodRegistry(ts storage.Storage, ds storage.Storage, l logrus.FieldLogger, options ...func(*MethodConfig)) (*MethodRegistry, error) {
-	rt, err := NewKVTable(ts)
-	if err != nil {
-		return nil, fmt.Errorf("Method new table: %w", err)
-	}
-	db, err := NewKVDocDB(ds)
-	if err != nil {
-		return nil, fmt.Errorf("Method new docdb: %w", err)
-	}
+var _ MethodManager = (*MethodRegistry)(nil)
 
-	conf := &MethodConfig{
-		Admin:         "did:bitxhub:relayroot:0x00000001",
-		Addr:          ".",
-		IsRoot:        true,
-		GenesisMetohd: "did:bitxhub:relayroot:.",
-		GenesisDoc: &MethodDoc{
-			BasicDoc: BasicDoc{
-				ID:   "did:bitxhub:relayroot:.",
-				Type: "method",
-				PublicKey: []PubKey{
-					{ID: "KEY#1",
-						Type:         "Secp256k1",
-						PublicKeyPem: "02b97c30de767f084ce3080168ee293053ba33b235d7116a3263d29f1450936b71"},
-				},
-				Controller: DID("did:bitxhub:relayroot:0x00000001"),
-				Authentication: []Auth{
-					{PublicKey: []string{"KEY#1"}},
+// MethodRegistry .
+type MethodRegistry struct {
+	mode          RegistryMode
+	isRoot        bool
+	admins        []DID
+	table         RegistryTable
+	docdb         DocDB
+	genesisMetohd DID
+	genesisDoc    DocOption
+	logger        logrus.FieldLogger
+}
+
+// NewMethodRegistry news a MethodRegistry
+func NewMethodRegistry(ts storage.Storage, l logrus.FieldLogger, options ...func(*MethodRegistry)) (*MethodRegistry, error) {
+	rt, _ := NewKVTable(ts)
+	db, _ := NewKVDocDB(nil)
+
+	mr := &MethodRegistry{ // default config
+		mode:          ExternalDocDB,
+		table:         rt,
+		docdb:         db,
+		logger:        l,
+		admins:        []DID{"did:bitxhub:relayroot:0x00000001"},
+		isRoot:        true,
+		genesisMetohd: "did:bitxhub:relayroot:.",
+		genesisDoc: DocOption{
+			Content: &MethodDoc{
+				BasicDoc: BasicDoc{
+					ID:   "did:bitxhub:relayroot:.",
+					Type: "method",
+					PublicKey: []PubKey{
+						{ID: "KEY#1",
+							Type:         "Secp256k1",
+							PublicKeyPem: "02b97c30de767f084ce3080168ee293053ba33b235d7116a3263d29f1450936b71"},
+					},
+					Controller: DID("did:bitxhub:relayroot:0x00000001"),
+					Authentication: []Auth{
+						{PublicKey: []string{"KEY#1"}},
+					},
 				},
 			},
 		},
 	}
+
 	for _, option := range options {
-		option(conf)
+		option(mr)
 	}
 
-	return &MethodRegistry{
-		table:  rt,
-		docdb:  db,
-		logger: l,
-		config: conf,
-		admins: []DID{conf.Admin},
-	}, nil
+	return mr, nil
+}
+
+// WithDocStorage .
+func WithDocStorage(ds storage.Storage) func(*MethodRegistry) {
+	return func(m *MethodRegistry) {
+		db, _ := NewKVDocDB(ds)
+		m.docdb = db
+		m.mode = InternalDocDB
+	}
 }
 
 // WithAdmin .
-func WithAdmin(a DID) func(*MethodConfig) {
-	return func(s *MethodConfig) {
-		s.Admin = a
+func WithAdmin(a DID) func(*MethodRegistry) {
+	return func(m *MethodRegistry) {
+		m.admins = []DID{a}
 	}
 }
 
 // WithGenesisMetohd .
-func WithGenesisMetohd(m DID) func(*MethodConfig) {
-	return func(s *MethodConfig) {
-		s.GenesisMetohd = m
+func WithGenesisMetohd(m DID) func(*MethodRegistry) {
+	return func(mr *MethodRegistry) {
+		mr.genesisMetohd = m
 	}
 }
 
 // WithGenesisDoc .
-func WithGenesisDoc(doc *MethodDoc) func(*MethodConfig) {
-	return func(s *MethodConfig) {
-		s.GenesisDoc = doc
+func WithGenesisDoc(doc *MethodDoc) func(*MethodRegistry) {
+	return func(mr *MethodRegistry) {
+		mr.genesisDoc = DocOption{Content: doc}
 	}
 }
 
 // SetupGenesis set up genesis to boot the whole methed system
-func (r *MethodRegistry) SetupGenesis() error {
-	if !r.config.IsRoot {
+func (r *MethodRegistry) SetupGenesis() error { // docOption DocOption
+	if !r.isRoot {
 		return fmt.Errorf("Method genesis registry not root")
 	}
-	if r.config.GenesisMetohd != r.config.GenesisDoc.ID {
+	if r.genesisMetohd != r.genesisDoc.Content.(*MethodDoc).ID {
 		return fmt.Errorf("Method genesis Method not matched with doc")
 	}
+
 	// register method
-	err := r.Apply(r.config.Admin, r.config.GenesisMetohd)
+	err := r.Apply(r.admins[0], r.genesisMetohd)
 	if err != nil {
-		return fmt.Errorf("Method genesis err: %w", err)
+		return fmt.Errorf("Method genesis apply err: %w", err)
 	}
 
-	err = r.AuditApply(r.config.GenesisMetohd, true)
+	err = r.AuditApply(r.genesisMetohd, true)
 	if err != nil {
-		return fmt.Errorf("Method genesis err: %w", err)
+		return fmt.Errorf("Method genesis audit err: %w", err)
 	}
-	r.logger.Info()
-	_, _, err = r.Register(r.config.GenesisDoc)
+
+	_, _, err = r.Register(r.genesisDoc)
 	if err != nil {
-		return fmt.Errorf("Method genesis err: %w", err)
+		return fmt.Errorf("Method genesis register err: %w", err)
 	}
-	// add admins did
-	r.AddAdmin(DID(r.config.Admin))
 
 	return nil
 }
 
 // GetSelfID .
 func (r *MethodRegistry) GetSelfID() DID {
-	return r.config.GenesisMetohd
+	return r.genesisMetohd
 }
 
 // GetAdmins .
@@ -201,7 +206,7 @@ func (r *MethodRegistry) Apply(caller DID, method DID) error {
 
 	status := r.getMethodStatus(method)
 	if status != Initial {
-		return fmt.Errorf("can not apply Method under status: %d", status)
+		return fmt.Errorf("can not apply %s under status: %s", method, status)
 	}
 	// creates item in table
 	err := r.table.CreateItem(
@@ -211,7 +216,7 @@ func (r *MethodRegistry) Apply(caller DID, method DID) error {
 				Status: ApplyAudit},
 			caller})
 	if err != nil {
-		return fmt.Errorf("apply Method on table: %w", err)
+		return fmt.Errorf("apply %s on table: %w", method, err)
 	}
 	return nil
 }
@@ -221,11 +226,11 @@ func (r *MethodRegistry) Apply(caller DID, method DID) error {
 func (r *MethodRegistry) AuditApply(method DID, result bool) error {
 	exist := r.HasMethod(method)
 	if exist == false {
-		return fmt.Errorf("auditapply Method %s not existed", method)
+		return fmt.Errorf("auditapply %s not existed", method)
 	}
 	status := r.getMethodStatus(method)
 	if !(status == ApplyAudit || status == ApplyFailed) {
-		return fmt.Errorf("can not auditapply under status: %d", status)
+		return fmt.Errorf("can not auditapply %s under status: %s", method, status)
 	}
 	var err error = nil
 	if result {
@@ -243,73 +248,71 @@ func (r *MethodRegistry) Synchronize(item *MethodItem) error {
 
 // Register ties method name to a method doc
 // ATN: only did who owns method-name should call this
-func (r *MethodRegistry) Register(doc *MethodDoc) (string, []byte, error) {
-	method := DID(doc.ID)
-	status := r.getMethodStatus(method)
-	if status != ApplySuccess {
-		return "", nil, fmt.Errorf("can not register under status, %d", status)
-	}
-
-	docBytes, err := doc.Marshal()
-	if err != nil {
-		return "", nil, fmt.Errorf("Method register doc marshal, %w", err)
-	}
-
-	docAddr, err := r.docdb.Create(doc)
-	if err != nil {
-		return "", nil, fmt.Errorf("Method register on docdb, %w", err)
-	}
-	docHash := sha256.Sum256(docBytes)
-	// update MethodRegistry table
-	item, err := r.table.GetItem(method, MethodTableType)
-	if err != nil {
-		return "", nil, fmt.Errorf("Method register table get, %w", err)
-	}
-	itemM := item.(*MethodItem)
-	itemM.Status = Normal
-	itemM.DocAddr = docAddr
-	itemM.DocHash = docHash[:]
-	err = r.table.UpdateItem(itemM)
-	if err != nil {
-		return docAddr, itemM.DocHash, fmt.Errorf("Method register table update: %w", err)
-	}
-	return docAddr, itemM.DocHash, nil
+func (r *MethodRegistry) Register(docOption DocOption) (string, []byte, error) { // doc *MethodDoc
+	return r.update(docOption, ApplySuccess, Normal)
 }
 
 // Update .
 // ATN: only did who owns method-name should call this.
-func (r *MethodRegistry) Update(doc *MethodDoc) (string, []byte, error) {
-	// check exist
-	method := DID(doc.ID)
-	status := r.getMethodStatus(method)
-	if status != Normal {
-		return "", nil, fmt.Errorf("can not update under status: %d", status)
-	}
+func (r *MethodRegistry) Update(docOption DocOption) (string, []byte, error) {
+	return r.update(docOption, Normal, Normal)
+}
 
-	docBytes, err := doc.Marshal()
-	if err != nil {
-		return "", nil, fmt.Errorf("Method update doc marshal: %w", err)
-	}
+// update with expected status
+func (r *MethodRegistry) update(docOption DocOption, expectedStatus StatusType, status StatusType) (string, []byte, error) {
+	var docAddr string
+	var docHash []byte
+	var method DID
+	if r.mode == InternalDocDB {
+		// check exist
+		doc := docOption.Content.(*MethodDoc)
+		method = doc.GetID()
+		status := r.getMethodStatus(method)
+		if status != expectedStatus {
+			return "", nil, fmt.Errorf("Method %s is under status: %s, expectd status: %s", method, status, expectedStatus)
+		}
 
-	docAddr, err := r.docdb.Update(doc)
-	if err != nil {
-		return "", nil, fmt.Errorf("Method update on docdb: %w", err)
+		docBytes, err := doc.Marshal()
+		if err != nil {
+			return "", nil, fmt.Errorf("doc marshal: %w ", err)
+		}
+
+		if r.docdb.Has(doc.GetID()) {
+			docAddr, err = r.docdb.Update(doc)
+		} else {
+			docAddr, err = r.docdb.Create(doc)
+		}
+
+		if err != nil {
+			return "", nil, fmt.Errorf("update docdb: %w ", err)
+		}
+
+		docHash32 := sha256.Sum256(docBytes)
+		docHash = docHash32[:]
+	} else {
+		method = docOption.ID
+		docAddr = docOption.Addr
+		docHash = docOption.Hash
+		status := r.getMethodStatus(method)
+		if status != expectedStatus {
+			return "", nil, fmt.Errorf("Method %s is under status: %s, expectd status: %s", method, status, expectedStatus)
+		}
 	}
-	docHash := sha256.Sum256(docBytes)
 
 	item, err := r.table.GetItem(method, MethodTableType)
 	if err != nil {
-		return docAddr, docHash[:], fmt.Errorf("Method update table get: %w", err)
+		return docAddr, docHash[:], fmt.Errorf("table get item: %w ", err)
 	}
 	itemM := item.(*MethodItem)
 	itemM.DocAddr = docAddr
-	itemM.DocHash = docHash[:]
+	itemM.DocHash = docHash
+	itemM.Status = status
 	err = r.table.UpdateItem(itemM)
 	if err != nil {
-		return docAddr, docHash[:], fmt.Errorf("Method update table update: %w", err)
+		return docAddr, docHash, fmt.Errorf("table update item: %w ", err)
 	}
 
-	return docAddr, docHash[:], nil
+	return docAddr, docHash, nil
 }
 
 // Audit .
@@ -317,7 +320,7 @@ func (r *MethodRegistry) Update(doc *MethodDoc) (string, []byte, error) {
 func (r *MethodRegistry) Audit(method DID, status StatusType) error {
 	exist := r.HasMethod(method)
 	if exist == false {
-		return fmt.Errorf("audit Method %s not existed", method)
+		return fmt.Errorf("audit %s not existed", method)
 	}
 	return r.auditStatus(method, status)
 }
@@ -327,7 +330,7 @@ func (r *MethodRegistry) Audit(method DID, status StatusType) error {
 func (r *MethodRegistry) Freeze(method DID) error {
 	exist := r.HasMethod(method)
 	if exist == false {
-		return fmt.Errorf("freeze Method %s not existed", method)
+		return fmt.Errorf("freeze %s not existed", method)
 	}
 	return r.auditStatus(method, Frozen)
 }
@@ -337,7 +340,7 @@ func (r *MethodRegistry) Freeze(method DID) error {
 func (r *MethodRegistry) UnFreeze(method DID) error {
 	exist := r.HasMethod(method)
 	if exist == false {
-		return fmt.Errorf("unfreeze Method %s not existed", method)
+		return fmt.Errorf("unfreeze %s not existed", method)
 	}
 
 	return r.auditStatus(method, Normal)
@@ -347,10 +350,15 @@ func (r *MethodRegistry) UnFreeze(method DID) error {
 func (r *MethodRegistry) Delete(method DID) error {
 	err := r.auditStatus(method, Initial)
 	if err != nil {
-		return fmt.Errorf("Method delete status aduit: %w", err)
+		return fmt.Errorf("Method delete: %w", err)
 	}
-	r.docdb.Delete(method)
+
 	r.table.DeleteItem(method)
+
+	if r.mode == InternalDocDB {
+		r.docdb.Delete(method)
+	}
+
 	return nil
 }
 
@@ -365,17 +373,16 @@ func (r *MethodRegistry) Resolve(method DID) (*MethodItem, *MethodDoc, bool, err
 		return nil, nil, false, fmt.Errorf("Method resolve table get: %w", err)
 	}
 	itemM := item.(*MethodItem)
-	doc, err := r.docdb.Get(method, MethodDocType)
-	if err != nil {
-		return itemM, nil, true, fmt.Errorf("Method resolve docdb get: %w", err)
+
+	docM := &MethodDoc{}
+	if r.mode == InternalDocDB {
+		doc, err := r.docdb.Get(method, MethodDocType)
+		if err != nil {
+			return itemM, nil, true, fmt.Errorf("Method resolve docdb get: %w", err)
+		}
+		docM = doc.(*MethodDoc)
 	}
-	docM := doc.(*MethodDoc)
 	return itemM, docM, true, nil
-}
-
-// MethodHasAccount checks whether account exists on the method blockchain
-func (r *MethodRegistry) MethodHasAccount(method string, account string) {
-
 }
 
 // HasMethod .
@@ -385,10 +392,13 @@ func (r *MethodRegistry) HasMethod(method DID) bool {
 }
 
 func (r *MethodRegistry) getMethodStatus(method DID) StatusType {
+	if !r.table.HasItem(method) {
+		return Initial
+	}
 	item, err := r.table.GetItem(method, MethodTableType)
 	if err != nil {
-		r.logger.Warn("method status get item:", err)
-		return Initial
+		r.logger.Error("method status get item:", err)
+		return Error
 	}
 	itemM := item.(*MethodItem)
 	return itemM.Status
@@ -398,13 +408,13 @@ func (r *MethodRegistry) getMethodStatus(method DID) StatusType {
 func (r *MethodRegistry) auditStatus(method DID, status StatusType) error {
 	item, err := r.table.GetItem(method, MethodTableType)
 	if err != nil {
-		return fmt.Errorf("Method aduit status table get: %w", err)
+		return fmt.Errorf("aduitstatus table get: %w", err)
 	}
 	itemM := item.(*MethodItem)
 	itemM.Status = status
 	err = r.table.UpdateItem(itemM)
 	if err != nil {
-		return fmt.Errorf("Method aduit status table update: %w", err)
+		return fmt.Errorf("aduitstatus table update: %w", err)
 	}
 	return nil
 }
